fix: Split payload tests
